+++ backend/src/main/resources/db/migration/V1__baseline.sql
@@ -6,7 +6,7 @@
 -- Authentication (compatible with Django's auth_user so we can carry passwords)
 -- ---------------------------------------------------------------------------
 CREATE TABLE auth_user (
-    id              BIGSERIAL PRIMARY KEY,
+    id              SERIAL PRIMARY KEY,
     password        VARCHAR(128)   NOT NULL,
     last_login      TIMESTAMPTZ    NULL,
     is_superuser    BOOLEAN        NOT NULL DEFAULT FALSE,
@@ -102,7 +102,7 @@ CREATE TABLE core_run (
 -- ---------------------------------------------------------------------------
 CREATE TABLE core_seenlisting (
     id          BIGSERIAL PRIMARY KEY,
-    user_id     BIGINT      NOT NULL REFERENCES auth_user(id)   ON DELETE CASCADE,
+    user_id     INTEGER     NOT NULL REFERENCES auth_user(id)   ON DELETE CASCADE,
     listing_id  BIGINT      NOT NULL REFERENCES core_joblisting(id) ON DELETE CASCADE,
     created_at  TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     CONSTRAINT unique_user_listing_seen UNIQUE (user_id, listing_id)
@@ -115,7 +115,7 @@ CREATE INDEX seen_user_listing_idx ON core_seenlisting(user_id, listing_id);
 -- ---------------------------------------------------------------------------
 CREATE TABLE core_savedview (
     id                  BIGSERIAL PRIMARY KEY,
-    user_id             BIGINT       NOT NULL REFERENCES auth_user(id) ON DELETE CASCADE,
+    user_id             INTEGER      NOT NULL REFERENCES auth_user(id) ON DELETE CASCADE,
     name                VARCHAR(255) NOT NULL,
     filter_expression   JSONB        NULL,
     columns             JSONB        NOT NULL,
